LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src-attr-elem/style-src-elem-allowed-attr-blocked-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'style-src-elem'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'style-src-attr'.


Harness Error (TIMEOUT), message = null

NOTRUN Should fire a security policy violation for the attribute
FAIL The attribute style should not be applied and the inline style should be applied assert_equals: expected 0 but got 10
